% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRW_SeqNet.R
\name{DRW_network}
\alias{DRW_network}
\title{Function for directed random walk algorithm}
\usage{
DRW_network(network, p0 = NULL, gamma = 0, tol = 1e-10)
}
\arguments{
\item{network}{Network to reflect the interactions among all predictors}

\item{p0}{The initial probability distribution over all nodes within the network}

\item{gamma}{The restart parameter in the directed random walk algorithm, the default value is 0}

\item{tol}{The tolerance parameter to reflect the precision of the estimated equilibrium probability distribution}
}
\value{
A vectror of estimated equilibrium probability distribution over all nodes in the network
}
\description{
Function for directed random walk algorithm
}
\examples{
\dontrun{
sim.data <- gen_data(num.sample = 500, num.var = 1000, scenario = 2, AveEffect = 2)
prob.topology <- DRW_network(network = sim.data$network)
}
}
