% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Generate synthetic gene network and RNA-Seq data for a given scenario}
\usage{
gen_data(
  num.sample = 1000,
  num.var = 1000,
  scenario = 2,
  AveEffect = 2,
  gamma = 0.3,
  num.rep = 100,
  seed = 11223344
)
}
\arguments{
\item{num.sample}{Number of samples of training set. The testing set has the same sample size.}

\item{num.var}{Number of genes simulated.}

\item{scenario}{A numeric ID to specify the simulation scenario. 1 for Null case,
2 for randomly distributed disease genes, 3 for one disease module without main disease gene,
4 for one disease module with one main disease gene, 5 for two disease modules without main disease genes,
6 for two disease modules with one main disease gene in each module}

\item{AveEffect}{The average effect size used in the logistic regression model to generate disease status}

\item{gamma}{The probability of re-start in the directed random walk algorithm}

\item{num.rep}{Number of repetitions}

\item{seed}{An integer number serving as the seed of the (pseudo-)random number generation}
}
\value{
A list of objects giving the synthetic datasets for a given scenario. The list consists of the following subjects:

* scenario The index of scenario
* num_sample Number of samples in training set and testing set
* num_var Number of genes simulated
* network The synthetic underlying gene network
* rna A list of 100 replications of RNA-Seq data with each replication consisting of both training and testing datasets
* disease_gene A vector of indices of genes corresponding to the disease genes of the simulation scenario
* main_disease_gene A vector of indices of genes corresponding to the main disease gene(s) of the scenario. When all disease genes having the same effect size, then all genes will be main disease genes. Otherwise, the index(es) of main disease gene(s) from disease module(s) is returned.
}
\description{
Generate synthetic gene network and RNA-Seq data for a given scenario
}
\examples{
\dontrun{
sim.data <- gen_data(num.sample = 500, num.var = 1000, scenario = 2, AveEffect = 2)
}
}
