% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_ladder.R
\name{gen_ladder}
\alias{gen_ladder}
\title{Function to generate a grid of integers}
\usage{
gen_ladder(Ntotal, ratio = 0.9, Nmin = 5)
}
\arguments{
\item{Ntotal}{Total number of predictors}

\item{ratio}{Rate for decrease the grid}

\item{Nmin}{The minimum number of the grid of integers}
}
\value{
A vector of integers in decreasing order starting from Ntotal and ending at Nmin.
}
\description{
Function to generate a grid of integers
}
\examples{
ladder <- gen_ladder(Ntotal = 1000, ratio = 0.9, Nmin = 20)

}
