% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weight.R
\name{get_weight}
\alias{get_weight}
\title{Function for constructing sampling probability of predictors used in the RF construction}
\usage{
get_weight(method = "uniform", x, y, network, gamma = 0.3)
}
\arguments{
\item{method}{Method for constructing sampling probability of predictors, currently supported methods are
* \"uniform\": all predictors have equal probability to be sampled which leads to the standard RF, default option
* \"marginal-p\": sampling probability derived from p-values of marginal two-sample t-test per gene
* \"topology\": sampling probability derived from network structure to only reflect the network topology
* \"network-p\": sampling probability derived from both network structure and marginal p-values
* \"network-q\": sampling probability derived from both network structure and FDR-adjusted marginal p-values}

\item{x}{Matrix of predictors}

\item{y}{Binary response vector}

\item{network}{Network to reflect the interactions among all predictors}

\item{gamma}{The restart parameter in the directed random walk algorithm, the default value is 0.3}
}
\value{
A vectror of sampling probability of all predictors used in the RF construction at each node
}
\description{
Function for constructing sampling probability of predictors used in the RF construction
}
\examples{
\dontrun{
sim.data <- gen_data(num.sample = 500, num.var = 1000, scenario = 2, AveEffect = 2)
Wt.marginal.p <- get_weight(method = "marginal-p",
                 x = sim.data$rna[[1]]$x_train,
                 y = sim.data$rna[[1]]$y_train)
Wt.topology <- get_weight(method = "topology", network = sim.data$network)
}
}
