% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkRF_binary.R
\name{networkRF_binary}
\alias{networkRF_binary}
\title{Construction of network-guided random forests for binary classification problem}
\usage{
networkRF_binary(x, y, Wt, min.num.gene, num.trees, ...)
}
\arguments{
\item{x}{Matrix of predictors}

\item{y}{Binary response vector}

\item{Wt}{Sampling probability of predictors used in the RF construction at each node}

\item{min.num.gene}{The number of genes retained at the end of the variable selection procedure}

\item{num.trees}{Number of trees in the RF construction}

\item{...}{Any other arguments allowed in the ranger function}
}
\value{
A list of objects giving the results of variable selection based on network-guided RF. The list consists of the following subjects:

* networkRF_model The final random forest predictive model constructed with top min.num.gene number of genes
* geneSelected A matrix recording the top selected genes at each step of the feature elimination procedure
* OOB.err A vector recording the OOB prediction error estimates for each RF predictive model constructed during the feature elimination procedure
* Wt Sampling probability of all predictors used in the RF construction at each node
}
\description{
Construction of network-guided random forests for binary classification problem
}
\examples{
\dontrun{
sim.data <- gen_data(num.sample = 500, num.var = 1000, scenario = 2, AveEffect = 2)
res.one.rep <- networkRF_binary(x = sim.data$rna[[1]]$x_train,
                                y = sim.data$rna[[1]]$y_train,
                                Wt = rep(1, num.var),
                                min.num.gene = 20,
                                num.trees = 2000)
}
}
