% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_accuracy.R
\name{pred_accuracy}
\alias{pred_accuracy}
\title{Function for prediction on testing data and get misclassification rate}
\usage{
pred_accuracy(xtest, ytest, rf.model)
}
\arguments{
\item{xtest}{Predictors of testing dataset}

\item{ytest}{Binary response vector of testing dataset}

\item{rf.model}{A random forest predictive model}
}
\value{
err A scalar of misclassification error of the rf.model on given testing dataset
}
\description{
Function for prediction on testing data and get misclassification rate
}
\examples{
sim.data <- gen_data(num.sample = 500, num.var = 1000, scenario = 2, AveEffect = 2)
res.one.rep <- networkRF(x = sim.data$rna[[1]]$x_train,
                         y = sim.data$rna[[1]]$y_train,
                         network = sim.data$network,
                         method = "topology",
                         min.num.gene = 20,
                         num.trees = 2000)
pred.accuracy <- pred_accuracy(x = sim.data$rna[[1]]$x_test,
                               y = sim.data$rna[[1]]$y_test,
                               rf.model = res.one.rep$networkRF_model)

}
